\name{dsadmin.get_methods}
\alias{dsadmin.get_methods}
\title{Get Datashield Methods}
\usage{
  dsadmin.get_methods(opal, type = "aggregate")
}
\arguments{
  \item{opal}{Opal object or list of opal objects.}

  \item{type}{Type of the method: "aggregate" (default) or
  "assign"}
}
\description{
  Get Datashield methods from Opal(s).
}

