\name{dsadmin.install_datashield}
\alias{dsadmin.install_datashield}
\title{Install a package from Datashield public source repository on GitHub.}
\usage{
  dsadmin.install_datashield(opal, pkg, ref = "master")
}
\arguments{
  \item{opal}{Opal object or list of opal objects.}

  \item{pkg}{Package name.}

  \item{ref}{Desired git reference. Could be a commit, tag,
  or branch name. Defaults to "master".}
}
\description{
  Install a package from Datashield public source
  repository on GitHub.
}

