\name{oadmin.install_package}
\alias{oadmin.install_package}
\title{Install package if not already available.}
\usage{
  oadmin.install_package(opal, pkg, repos = NULL)
}
\arguments{
  \item{opal}{Opal object or list of opal objects.}

  \item{pkg}{Package name.}

  \item{repos}{Character vector, the base URLs of the
  repositories to use.}
}
\description{
  Install package if not already available.
}

