\name{oadmin.installed_package}
\alias{oadmin.installed_package}
\title{Check if a package is installed.}
\usage{
  oadmin.installed_package(opal, pkg)
}
\arguments{
  \item{opal}{Opal object or list of opal objects.}

  \item{pkg}{Package name.}
}
\description{
  Check if a package is installed.
}

