\name{opal.install_datashield}
\alias{opal.install_datashield}
\title{Install a package from Datashield source repository on GitHub.}
\usage{
  opal.install_datashield(opal, pkg, ref = "master")
}
\arguments{
  \item{opal}{Opal object.}

  \item{pkg}{Package name.}

  \item{ref}{Desired git reference. Could be a commit, tag,
  or branch name. Defaults to "master".}
}
\description{
  Install a package from Datashield source repository on
  GitHub.
}

