\name{opal.package_description}
\alias{opal.package_description}
\title{Get package description from Opal(s).}
\usage{
  opal.package_description(opal, pkg)
}
\arguments{
  \item{opal}{Opal object or list of opal objects.}

  \item{pkg}{Package name.}
}
\description{
  Get package description from Opal(s).
}

