\name{opal.install_package}
\alias{opal.install_package}
\title{Install package if not already available.}
\usage{
  opal.install_package(opal, pkg)
}
\arguments{
  \item{opal}{Opal object or list of opal objects.}

  \item{pkg}{Package name.}
}
\description{
  Install package if not already available.
}

